Import des d'Untis compatible.
Opció per importar des d'Untis sobre un horari ja existent.
Reubicades algunes funcions.
Corregit error als migrate.
Actualitzada documentació.
