+++ docs/Wiki/manual-de-uso/carga-inicial-de-datos/carga-de-horarios-aulas-y-asignaturas.md
@@ -1,6 +1,10 @@
 # Carga de Horarios, Aulas y Asignaturas
 
-Lo único que queda por importar son los horarios, Aulas y Materias para ello utilizaremos de nuevo el fichero proporcionado por Kronowin.
+Lo único que queda por importar son los horarios, Aulas y Materias.
+
+Con Kronowin utilizaremos de nuevo el fichero proporcionado por Kronowin.
+
+Con Untis este paso ya se ha realizado, conjuntamente con profesorado y grupos, al realizar la importación del fichero xml
 
 Volvemos otra vez a [\[URL\_DJANGO\]/extKronowin/sincronitzaKronowin/](https://djau.local/extKronowin/sincronitzaKronowin/) y cargamos el mismo fichero de siempre, la salida deberá ser la siguiente:
